:sparkles: [Feat]: 회원, 게임스타일, 챔피언, 게시판 관련 JPA Entity 폴더 구조 변경
