+++ src/main/java/com/gamegoo/domain/gamestyle/MemberGameStyle.java
@@ -1,24 +1,26 @@
-package com.gamegoo.domain;
+package com.gamegoo.domain.gamestyle;
 
+import com.gamegoo.domain.Member;
 import lombok.Getter;
 import lombok.Setter;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
 
+@Entity
+@Table(name = "MemberGameStyle")
 @Getter
 @Setter
-@Entity
-@Table(name = "MemberChampion")
-public class MemberChampion {
+public class MemberGameStyle {
+
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    @Column(name = "member_champion_id", nullable = false)
+    @Column(name = "member_gamestyle_id", nullable = false)
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "champion_id", nullable = false)
-    private Champion champion;
+    @JoinColumn(name = "gamestyle_id", nullable = false)
+    private GameStyle gameStyle;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id", nullable = false)
